+++ games/nes/ng2/gameInstance.hpp
@@ -96,7 +96,7 @@ class GameInstance : public GameInstanceBase
   double ninjaBossDistance;
 
   GameInstance(EmuInstance* emu, const nlohmann::json& config);
-  uint64_t computeHash() const override;
+  uint128_t computeHash() const override;
   void updateDerivedValues() override;
   std::vector<std::string> getPossibleMoves() const override;
   magnetSet_t getMagnetValues(const bool* rulesStatus) const;
